+++ src/app/components/series-chart.tsx
@@ -2,6 +2,7 @@
 
 import { Typography } from "@/components/typography";
 import { AreaChart } from "@tremor/react";
+import Header from "./header";
 
 const formatCurrency = (value: number) => {
   return `$${value}`
@@ -25,26 +26,14 @@ export const SeriesChart = ({
   const currentValue = 100
 
   return (
-    <div>
-      <h2 className="scroll-m-20 text-3xl font-semibold tracking-tight first:mt-0">
-        {formatCurrency(currentValue)}
-      </h2>
-      <p className="leading-7 text-muted-foreground">
-        {formatCurrency(currentValue)}
-        <span>
-          {'against S&P500'}
-        </span>
-      </p>
-      <div className="h-72">
-        <AreaChart
-
-          data={data}
-          index="date"
-          categories={["index", "asset"]}
-          colors={["gray", "green"]}
-          valueFormatter={formatPoint}
-        />
-      </div>
+    <div className="w-full h-72">
+      <AreaChart
+        data={data}
+        index="date"
+        categories={["index", "asset"]}
+        colors={["gray", "green"]}
+        valueFormatter={formatPoint}
+      />
     </div>
   )
