+++ src/components/date-picker.tsx
@@ -7,17 +7,18 @@ import { CalendarIcon } from "@radix-ui/react-icons"
 import { cn } from "@/lib/utils"
 import { Button } from "@/components/button"
 
-function DatePickerTrigger({ value }: { value: Date | undefined }) {
+function DatePickerTrigger({ value, ...props }: { value: Date | undefined }) {
   return (
     <Button
-      variant={"outline"}
+      variant="outline"
       className={cn(
-        "w-[280px] justify-start text-left font-normal",
+        "w-[200px] justify-start text-left font-normal",
         !value && "text-muted-foreground"
       )}
+      {...props}
     >
       <CalendarIcon className="mr-2 h-4 w-4" />
-      {value ? format(value, "PPP") : <span>{'Pick a date'}</span>}
+      {value ? format(value, "PP") : <span>{'Pick a date'}</span>}
     </Button>
   )
 }
