+++ src/styles/globals.css
@@ -38,7 +38,6 @@
     --input: 214.3 31.8% 91.4%;
     --ring: 222.2 84% 4.9%;
     --radius: 0.3rem;
-    --text-muted-foreground: red;
   }
  
   .dark {
@@ -61,7 +60,6 @@
     --border: 217.2 32.6% 17.5%;
     --input: 217.2 32.6% 17.5%;
     --ring: 212.7 26.8% 83.9;
-    --text-muted-foreground: red;
   }
 }
